Build the order pipeline: checkout, payment, fulfilment, and the event glue
Schema
- V1: five roles (ADMIN, MODERATOR, PICKER, DRIVER, CUSTOMER); one default
  address per user, no billing address (it belongs to the card)
- V2: grocery catalogue — sell_by/price_by, catch-weight bands, pick
  tolerance, net content, per-product tax rate, storage type, age restriction
- V3/V4/V5: numeric(12,3) quantities throughout, for weighed goods
- V5: authorise-then-capture. PLACED → AUTHORIZED → PROCESSING →
  AWAITING_PAYMENT → PAID; three amounts (estimate, hold, capture) with the
  capture bounded by the hold, not the estimate; delivery speed and 2h window;
  48h booking horizon; delivery address by reference
- V6: billing accounts, off-session methods with SetupIntent evidence,
  webhook inbox deduped on (provider, event id); retry schedule dropped
- V9: fulfilment — picking tasks, per-line verdicts with the substitution
  conversation, deliveries with copied window and age check, append-only
  attempts
- Seed: twelve grocery aisles, 48 products, ten weight-priced

Modules
- inventory, order, payment, fulfilment built from empty; StockLedger
  redesigned around movements — there are no reservations
- catalog and cart carry the weight model; Money and the contracts take
  BigDecimal; VariantSummary owns the estimate/ceiling arithmetic
- CheckoutService: place, hold the card synchronously while the shopper is
  present, move. Three commits by design — Stripe is never called inside a
  transaction
- Stripe via stripe-java 29 with manual capture; the webhook inbox records
  first and acts second, and the 3-D Secure path closes through
  PaymentAuthorised
- settlePick: a weighed line is billed at no more than the ceiling it was
  quoted, a substitute at no more than the line it replaced — which is what
  guarantees capture ≤ hold
- Cancellation releases the hold and unwinds a pick by storage type

Structure
- shared closed with six @NamedInterface doors; the build immediately caught
  a helper leaking out of shared.config
- jobs/ renamed scheduler/ (the clock); queue jobs live in the module they
  serve; top-level jobs/ reserved for cross-module work
- @Externalized events over the transactional outbox to RabbitMQ;
  @ApplicationModuleListener for status moves; the order module is the only
  writer of orders.status
- OrderStream SSE fan-out over the same exchange, one exclusive queue per
  web instance
- Five scheduler workers: outbox drain, publication resubmit, proposal
  timeout, webhook drain + alarm, stale-hold alarm
- Three runtime roles from one image: app / worker / scheduler, chosen by
  SCHEDULER_ENABLED and JOBS_ENABLED; infrastructure supplied by .env

Endpoints
- billing: setup, methods, default, remove
- checkout: slots, place
- orders: list, view, cancel, live stream, substitute proposals and answers
- webhooks: Stripe, permitAll — was behind auth
- staff paths gated by the role the token already carries

Fixes
- PaymentService rolled back the FAILED row on a decline (noRollbackFor)
- payment_method.attached now fills brand and last4; register() merges
- shared.infra.Queues so modules never import shared.config

Frontend: cart, profile sections, animations, customer dashboard

+++ backend/src/main/resources/db/migration/V3__inventory.sql
@@ -11,8 +11,10 @@
 CREATE TABLE inventory_items (
     variant_id         uuid PRIMARY KEY
         REFERENCES product_variants(id) ON DELETE CASCADE,   -- crosses into: catalog
-    quantity_on_hand   integer     NOT NULL DEFAULT 0 CHECK (quantity_on_hand >= 0),
-    quantity_reserved  integer     NOT NULL DEFAULT 0 CHECK (quantity_reserved >= 0),
+    -- Numeric, not integer: a shelf holds 12.400 kg of loose apples as readily
+    -- as it holds 12 tins, and a deli pick takes 0.540 of a pound off it.
+    quantity_on_hand   numeric(12,3) NOT NULL DEFAULT 0 CHECK (quantity_on_hand >= 0),
+    quantity_reserved  numeric(12,3) NOT NULL DEFAULT 0 CHECK (quantity_reserved >= 0),
     reorder_level      integer     NOT NULL DEFAULT 0 CHECK (reorder_level >= 0),
     -- allow selling into negative stock (backorder) for this variant
     allow_backorder    boolean     NOT NULL DEFAULT false,
@@ -40,7 +42,7 @@ FROM   inventory_items;
 CREATE TABLE inventory_reservations (
     id          uuid PRIMARY KEY DEFAULT gen_random_uuid(),
     variant_id  uuid        NOT NULL REFERENCES product_variants(id) ON DELETE CASCADE,
-    quantity    integer     NOT NULL CHECK (quantity > 0),
+    quantity    numeric(12,3) NOT NULL CHECK (quantity > 0),
     -- what the hold is for: a cart line first, an order once placed
     owner_type  varchar(16) NOT NULL CHECK (owner_type IN ('CART', 'ORDER')),
     owner_id    uuid        NOT NULL,
@@ -63,7 +65,7 @@ CREATE TABLE stock_movements (
     id            uuid PRIMARY KEY DEFAULT gen_random_uuid(),
     variant_id    uuid        NOT NULL REFERENCES product_variants(id) ON DELETE RESTRICT,
     -- signed: positive on restock/return, negative on fulfilment
-    quantity_delta integer    NOT NULL CHECK (quantity_delta <> 0),
+    quantity_delta numeric(12,3) NOT NULL CHECK (quantity_delta <> 0),
     reason        varchar(32) NOT NULL
         CHECK (reason IN ('RESTOCK', 'FULFILMENT', 'RETURN', 'ADJUSTMENT', 'DAMAGE')),
     reference_type varchar(16),
